==> 01_extract_and_clean_loa_data.ipynb <==
# Extração e limpeza dos dados da LOA

**objetivo**: identificar as emendas propostas pela legislatura 2021-2024 e cruzar com o vereador/partido que propôs.

Scripts para extrair dados das LOAs para os anos de 2022 a 2025. Essas LOAs correspondem à legislatura de 2021-2024, eleita nas eleições de 2020.

Serão extraídas também a relação de vereadores que foram proponentes de cada emenda.

## extração de dados
Extraindo os arquivos da LOA

import pandas as pd
from typing import Optional
import plotly.graph_objects as go
import os

from utils.download_file import download_csv_as_dataframe

from config import DATA_FOLDER, GRAFICOS_FOLDER 


pd.set_option('display.max_columns', None)

class DownloadLoa:

    base_url = 'https://orcamento.sf.prefeitura.sp.gov.br/orcamento/uploads'

    files = {

        2025 : 'LOA.25_VersionamentoDespesa.csv',
        2024 : 'OrcamentoAprovado_2024.csv',
        2023 : 'OrcamentoAprovado_2023.csv',
        2022 : 'OrcamentoAprovado_2022.csv'
    }

    def __build_url(self, year: int) -> str:
        """
        Builds the complete URL for downloading the LOA file for a given year.

        Args:
            year (int): The year for which the LOA file URL is to be built.

        Returns:
            str: The complete URL as a string.
        """
        endpoint = f'{year}/{self.files[year]}'
        return self.base_url + '/' + endpoint

    def __download(self, year: int, **read_kwargs)->pd.DataFrame:
        """
        Downloads the LOA file for a given year and returns it as a pandas DataFrame.

        Args:
            year (int): The year for which the LOA file is to be downloaded.

        Returns:
            pd.DataFrame: The downloaded LOA file as a pandas DataFrame.
        """



        url = self.__build_url(year)
        df = download_csv_as_dataframe(url, **read_kwargs)
        
        return df

    def __concat_year_dfs(self, year_list:[str], **read_kwargs)->pd.DataFrame:
        """
        Concatenates DataFrames for the specified years into a single DataFrame.

        Args:
            year_list (list): List of years for which DataFrames are to be concatenated.

        Returns:
            pd.DataFrame: The concatenated DataFrame.
        """
        dfs = [self.__download(year, **read_kwargs) for year in year_list]

        return pd.concat(dfs, ignore_index=True)

    def __describe_concated_df(self, df: pd.DataFrame) -> None:
        """
        Prints the shape, columns, and missing values for the given DataFrame.

        Args:
            df (pd.DataFrame): The DataFrame to describe.
        """
        print("Shape of DataFrame:", df.shape)
        print("Columns in DataFrame:", df.columns.tolist())
        print("Missing values per column:")
        print(df.isnull().sum())


    def __call__(self, year_list: Optional[[str]]=None, **read_kwargs)->pd.DataFrame:
        """
        Downloads and concatenates LOA files for the specified years.

        Args:
            year_list (list): List of years for which LOA files are to be downloaded.

        Returns:
            pd.DataFrame: The concatenated DataFrame containing the LOA data.
        """
        if year_list is None:
            year_list = self.files.keys()

        df = self.__concat_year_dfs(year_list, **read_kwargs)
        self.__describe_concated_df(df)

        return df


download_loa = DownloadLoa()

## limpeza de dados

padronizar colunas

df_loa = download_loa(encoding='latin-1', sep=';')

missing_values_count = df_loa.groupby("EXERCICIO").apply(lambda group: group.isnull().sum())


missing_values_percent = missing_values_count/df_loa.groupby("EXERCICIO").count()

missing_values_percent

cols = [col for col in missing_values_count.columns if missing_values_percent[col].sum() > 0 or missing_values_percent[col].isnull().all()]

missing_values_percent[cols]

cols_year = {year: set([col for col in df_loa if missing_values_percent.loc[year, col] < 100
            or pd.isnull(missing_values_percent.loc[year, col])
            ]) for year in range(2022,2026)}

for curr_year, cols_that_year in cols_year.items():
    
    cols_other_years=set()
    for year in cols_year:
        if year!= curr_year: cols_other_years.update(cols_year[year])

    print(f'Colunas especificas do ano: {year}', cols_that_year-cols_other_years)

cols_comuns = []
for col in missing_values_percent.columns:
    for year, cols_that_year in cols_year.items():
        if col not in cols_that_year:
            print(f'Coluna {col} não está presente no ano {year}')
            break
    else:
        print(f'Coluna {col} está presente em todos os anos')
        cols_comuns.append(col)

Conclui-se que há uma alteração no schema dos dados a partir de 2024, sendo que 2023 tem algumas das colunas que passam a aparecer em 2024. 2022 tem apenas uma coluna, "DOTACAO", que não aparece nos demais. A partir de 2024, as alterações nas colunas são conhecidas, pois devem-se ao novo padrão para registro detalhado das fontes das dotações.

cols_comuns = set(cols_comuns)

cols_year[2025].difference(cols_comuns)

cols_year[2024].difference(cols_comuns)

cols_year[2023].difference(cols_comuns)

cols_year[2022].difference(cols_comuns)

cols_comuns

Precisamos unificar e padronizar as colunas, mas vamos fazê-lo apenas para aquelas que fazem sentido para a análise, tentando manter principalmente as colunas que são comuns para todos os anos

colunas_padronizado = {
    'EXERCICIO' : 'exercicio',
    'COD_NRO_EMND_DOTA' : 'cod_emenda',
    'ADMINISTRACAO' : 'direta_ou_indireta',
    'ORGAO' : 'cod_orgao',
    'UNIDADE' : 'cod_unidade',
    'PROGRAMA' : 'cod_programa',
    'FUNCAO' : 'cod_funcao',
    'SUBFUNCAO' : 'cod_subfuncao',
    'PROJ_ATIV'  : 'cod_proj_atividade',
    'FONTE' : 'cod_fonte',
    'DESC_ORGAO' : 'desc_orgao',
    'DESC_UNIDADE' : 'desc_unidade',
    'DESC_PROGRAMA' : 'desc_programa',
    'DESC_FUNCAO' : 'desc_funcao',
    'DESC_SUBFUNCAO' : 'desc_subfuncao',
    'DESC_PROJ_ATIV' : 'desc_proj_atividade',
    'DESC_FONTE' : 'desc_fonte',
    'SALDO_ORÇ' : 'valor_orcado'
    }

cols_interesse = [col for col in colunas_padronizado.keys()]

df_loa = df_loa[cols_interesse].copy()
df_loa.rename(columns=colunas_padronizado, inplace=True)

df_loa.head()

### criando colunas de index

for col in ['exercicio', 'cod_orgao', 'cod_unidade', 'cod_proj_atividade']:
    print(col, df_loa[col].isnull().any())

def fill_zeros(df:pd.DataFrame, col:str, num_zeros:int)->None:
    """
    Fills NaN values in the specified column of the DataFrame with zeros.

    Args:
        df (pd.DataFrame): The DataFrame to modify.
        col (str): The column name to fill with zeros.
    """
    df[col] = df[col].astype(int).astype(str)
    df[col] = df[col].str.zfill(num_zeros)

    print(col, df[col].str.len().unique())

fill_zeros(df_loa, 'exercicio', 4)
fill_zeros(df_loa, 'cod_orgao', 2)
fill_zeros(df_loa, 'cod_unidade', 2)
fill_zeros(df_loa, 'cod_proj_atividade', 4)

df_loa['index'] = df_loa[['exercicio', 'cod_orgao', 'cod_unidade', 'cod_proj_atividade']].astype(str).agg('.'.join, axis=1)

df_loa['index'].sample(3)

O índice tem duplicadas, mas tudo bem. Ele não pode ter duplicadas na outra base, que identifica os parlamentares.

df_loa['index'].duplicated().any()

df_loa['index'].duplicated().sum()

### criando coluna de tratamento e avaliando

df_loa['is_emenda'] = df_loa['cod_emenda'].notnull()

df_loa['is_emenda'].mean()

df_loa['is_emenda'].sum()

emenda_count_per_year = df_loa.groupby('exercicio')['is_emenda'].sum()
emenda_count_per_year

emenda_percent_per_year = df_loa.groupby('exercicio')['is_emenda'].mean()
emenda_percent_per_year


# Dados para o gráfico
anos = emenda_count_per_year.index.astype(int)
quantidade_emendas = emenda_count_per_year.values
percentual_emendas = emenda_percent_per_year.values * 100  # Converter para porcentagem

# Criar o gráfico
fig = go.Figure()

# Adicionar barras para a quantidade de emendas
fig.add_trace(go.Bar(
    x=anos,
    y=quantidade_emendas,
    name='Quantidade de Emendas',
    marker_color='blue'
))

# Adicionar anotações para o percentual em cada barra
i = 0
for ano, percentual in zip(anos, percentual_emendas):
    fig.add_annotation(
        x=ano,
        y=emenda_count_per_year[i],
        text=f'{percentual:.2f}%',
        showarrow=False,
        font=dict(size=12, color='black'),
        yshift=10  # Ajustar a posição vertical da anotação
    )
    i+=1

# Configurar layout
fig.update_layout(
    title='Evolução da Quantidade e Percentual de Emendas por Ano',
    xaxis = dict(
        title='Ano',
        tickvals=anos,  # Define os valores dos ticks no eixo x
        ticktext=[str(ano) for ano in anos]  # Define os rótulos dos ticks
    ),
    yaxis_title='Valores',
    barmode='group',
    legend_title='Métricas',
    template='plotly_white'
)

fig.write_image(os.path.join(GRAFICOS_FOLDER, 'qtd_emendas_ano.png'))

# Mostrar o gráfico
fig.show()

### padronizada final

df_loa.dtypes

for col in df_loa.columns:
    if col.startswith('cod'):
        df_loa[col] = df_loa[col].fillna('').astype(str)

df_loa.dtypes

df_loa.describe()

df_loa.to_csv(os.path.join(DATA_FOLDER, 'df_loa.csv'), index=False, sep=';')

!which python
