==> 03_join_regionalizacao_loa.ipynb <==
## Join Regionalização LOA

Neste notebook fazemos o join dos datasets gerados nos notebooks 1 e 2.

Esse join é realizado apenas para as dotações da LOA que representam emendas de vereador.

import pandas as pd
import os

from config import DATA_FOLDER

## carregar dados

df_loa = pd.read_csv(os.path.join(DATA_FOLDER, 'df_loa.csv'), low_memory=False, sep=';')
df_da = pd.read_csv(os.path.join(DATA_FOLDER, 'df_da.csv'), low_memory=False, sep=';')

df_loa.head()

## Selecionando apenas emendas

Ao todo são mil emendas nos 4 anos da legislatura. O  dataset que estamos construindo trabalha apenas com elas.

df_loa['is_emenda'].mean()

df_loa['is_emenda'].sum()

df_loa = df_loa[df_loa['is_emenda']].copy().reset_index(drop=True)

df_loa['is_emenda'].all()

df_loa.drop('is_emenda', axis=1, inplace=True)

df_loa['cod_emenda'].isnull().any()

df_loa['cod_emenda'].sample(5)

#posso transformar em int agora porque não tem mais NAN
df_loa['cod_emenda'] = df_loa['cod_emenda'].astype(int)

df_loa.groupby('exercicio').count()['cod_emenda']

## Join dataframes

Agora vamos fazer join nos dataframes com base nos indices que foram criados.

Os dados das emendas (df_loa) ficam na esquerda, e é um left join. Ou seja: estamos buscando as regionalizações das emendas.

df_loa.dtypes

df_da.dtypes

df_merge = pd.merge(df_loa, df_da, how='left', on='index', suffixes=('', '_da'))
df_merge.head()

df_merge['desc_regiao'].isnull().any()

df_merge['desc_regiao'].value_counts()

df_merge.to_csv(os.path.join(DATA_FOLDER, 'df_merge_loa_da.csv'), sep=';', index=False)
