==> 04_extract_and_clean_emendas_data.ipynb <==
# Extração e limpeza dos dados de emendas

Neste notebook extraímos e limpamos os dados da relação de emendas por vereador.

Como no notebook anterior, são utilizadas as relações de emendas por vereador para os anos de 2022 a 2025 (ou seja, referentes às LOAs da legislatura eleita em 2020).

import requests
import xml.etree.ElementTree as ET
import json
import pandas as pd
import plotly.express as px
import plotly.graph_objects as go
from typing import Optional
import os

from config import GRAFICOS_FOLDER, DATA_FOLDER


## Extração de dados

class ExtractEmendasDataYear:

    base_url = 'http://saeows.saopaulo.sp.leg.br/Servico/'

    def __init__(self, service:str, endpoint:str)->None:
        """
        Initialize the ExtractEmendasDataYear class.

        Parameters:
        base_url (str): The base URL for the API.
        endpoint (str): The endpoint for the API.
        """

        if service.startswith('/'):
            service = service[1:]

        if not service.endswith('/'):
            service += '/'

        if not endpoint.startswith('/'):
            endpoint = '/' + endpoint

        if endpoint.endswith('/'):
            endpoint = endpoint[:-1]

        self.base_url = self.base_url + service
        self.endpoint = endpoint

    def __build_url(self, ano: int) -> str:
        """
        Build the URL for the API request.
        """
        return f'{self.base_url}{self.endpoint}?exercicio={ano}'

    def __get_data(self, ano: int) -> str:
        """
        Get data from the API for a specific year.

        Parameters:
        ano (int): The year for which data is to be fetched.

        Returns:
        str: The response text (a XML with JSON data).
        """

        url = self.__build_url(ano)

        response = requests.get(url)
        if response.status_code == 200:
            return response.text
        else:
            raise RuntimeError(f"Failed to fetch data for year {ano}: {response.status_code}")

    def __build_xml_tree(self, xml_string:str)-> ET.Element:
        """
        Build an XML tree from a string.

        Parameters:
        xml_string (str): The XML string to be parsed.

        Returns:
        Element: The root element of the parsed XML tree.
        """
        return ET.fromstring(xml_string)

    def __parse_xml_to_json(self, xml_string: str) -> dict:
        """
        Parse the XML string and extract the JSON data.

        Parameters:
        xml_string (str): The XML string to be parsed.

        Returns:
        dict: The extracted JSON data.
        """
        root = self.__build_xml_tree(xml_string)
        json_str = root.text.strip()

        try:
            json_data = json.loads(json_str)
            return json_data

        except json.JSONDecodeError as e:
            raise RuntimeError(f"Erro ao decodificar JSON : {e}")

    def __json_to_dataframe(self, json_data:dict) -> pd.DataFrame:
        """
        Convert JSON data to a pandas DataFrame.

        Parameters:
        json_data (dict): The JSON data to be converted.

        Returns:
        DataFrame: The resulting DataFrame.
        """
        df = pd.json_normalize(json_data)
        return df

    def __pipeline(self, ano:int)->pd.DataFrame:
        """
        Execute the data extraction pipeline for a specific year.

        Parameters:
        ano (int): The year for which data is to be processed.

        Returns:
        DataFrame: The resulting DataFrame.
        """
        xml_string = self.__get_data(ano)
        json_data = self.__parse_xml_to_json(xml_string)
        df = self.__json_to_dataframe(json_data)

        return df

    def __call__(self, ano: int) -> pd.DataFrame:
        """
        Execute the data extraction pipeline for a specific year when the object is called.

        Parameters:
        ano (int): The year for which data is to be processed.

        Returns:
        DataFrame: The resulting DataFrame.
        """

        return self.__pipeline(ano)


ANOS = tuple(range(2022, 2026))


def load_emendas(fname:str, endpoint:str, service:str='IntegracaoPMSP.asmx', anos:[int]=ANOS)->pd.DataFrame:
    """
    Load emendas data for the specified years.

    Parameters:
    anos (list): List of years to load data for.

    Returns:
    DataFrame: The combined DataFrame containing data for all specified years.
    """

    extract_emenda_data = ExtractEmendasDataYear(
        service = service,
        endpoint = endpoint
        )

    fpath = os.path.join(DATA_FOLDER, fname)
    if os.path.exists(fpath):
        print('Loading already downloaded data')
        df = pd.read_csv(fpath, sep=';')
        return df

    dfs = []
    for ano in anos:
        print('Extracting data for year', ano)
        df = extract_emenda_data(ano)
        df['ano'] = ano
        dfs.append(df)

    df = pd.concat(dfs, ignore_index=True)
    df.to_csv(fpath, index=False, sep=';')

    return df

df_emendas_vereadores = load_emendas(fname='emendas_vereaadores.csv',endpoint='EmendasVereadoresJson')

df_emendas_vereadores['ano'].unique()

df_vereadores = load_emendas(fname='vereadores.csv',endpoint='VereadoresJson')

df_vereadores['ano'].unique()

df_partidos = load_emendas(fname='partidos.csv', endpoint='PartidosJson')

df_partidos['ano'].unique()

## Filtrando as colunas

Não vamos usar todas as colunas no estudo. Então agora filtramos os dataframes

df_emendas_vereadores.head()

as emendas de vereadores tem duplicadas. provavelmente por conta da data da proposição. vou ordenar pela data e pegar a ultima

df_emendas_vereadores[['Numero', 'Vereador', 'ano']].duplicated().any()

df_emendas_vereadores[['Numero', 'Vereador', 'ano']].duplicated().mean()

df_emendas_vereadores[['Numero', 'Vereador', 'ano', 'Motivo']].duplicated().mean()

df_emendas_vereadores['DataEmenda'] = pd.to_datetime(df_emendas_vereadores['DataEmenda'], format='%d/%m/%Y', errors='coerce')
df_emendas_vereadores.head()

df_emendas_vereadores['DataEmenda'].isna().sum()

#algumas emendas não possuem data. vamos colocar elas com uma data no futuro para que elas não sejam perdidas

df_emendas_vereadores[df_emendas_vereadores['DataEmenda'].isnull()]

emendas_sem_data = df_emendas_vereadores[df_emendas_vereadores['DataEmenda'].isnull()][['Numero', 'ano']] \
    .apply(lambda row: f"{row['Numero']}-{row['ano']}", axis=1).tolist()

print(len(emendas_sem_data))
emendas_sem_data = set(emendas_sem_data)

emendas_com_data = df_emendas_vereadores[df_emendas_vereadores['DataEmenda'].notnull()][['Numero', 'ano']] \
    .apply(lambda row: f"{row['Numero']}-{row['ano']}", axis=1).tolist()
print(len(emendas_com_data))
emendas_com_data = set(emendas_com_data)

emendas_sem_data_nenhuma = emendas_sem_data - emendas_com_data

len(emendas_sem_data_nenhuma)

def set_ano_emenda(row)->str:

    if pd.isnull(row['DataEmenda']) and \
        f"{row['Numero']}-{row['ano']}" in emendas_sem_data_nenhuma:
            return '31/12/2025'

    return row['DataEmenda']

df_emendas_vereadores['DataEmenda'] = df_emendas_vereadores.apply(set_ano_emenda,axis =1)

df_emendas_vereadores['DataEmenda'].isnull().any()

df_emendas_vereadores['DataEmenda'] = pd.to_datetime(df_emendas_vereadores['DataEmenda'], format='%d/%m/%Y', errors='coerce')

df_emendas_vereadores['DataEmenda'].max()

df_emendas_vereadores['DataEmenda'].min()

df_emendas_vereadores = df_emendas_vereadores.sort_values(by='DataEmenda', ascending=True).reset_index(drop=True)

Vamos ficar com a ultima propositura de cada emenda

df_emendas_vereadores.drop_duplicates(['Numero', 'Vereador', 'ano'], inplace=True, keep='last')

df_emendas_vereadores = df_emendas_vereadores[['Numero', 'Vereador', 'ano', 'Motivo']].copy()

df_emendas_vereadores.rename(columns={'Numero': 'num_emenda', 'Vereador': 'cod_vereador', 'Motivo' : 'motivo'}, inplace=True)

df_emendas_vereadores.head()

df_emendas_vereadores.duplicated().any()

df_vereadores.head()

df_vereadores = df_vereadores[['Numero', 'Apelido', 'Partido', 'ano']].copy()
df_vereadores.rename(columns={'Numero': 'cod_vereador', 'Apelido': 'apelido', 'Partido' : 'cod_partido'}, inplace=True)

df_vereadores.head()

df_vereadores.duplicated().any()

df_partidos.head()

df_partidos = df_partidos[['Numero', 'Sigla', 'Nome', 'ano']].copy()

df_partidos.rename(columns={'Numero': 'cod_partido', 'Sigla': 'sigla', 'Nome' : 'nome_partido'}, inplace=True)

df_partidos.duplicated().any()

## Joins


Agora vamos unificar os datasets em um único

### Join vereadores

Unir vereadores e partidos.

df_vereadores['cod_partido'].isnull().any()

df_vereadores['cod_partido'].sample(3)  

df_partidos['cod_partido'].isnull().any()

df_partidos['cod_partido'].sample(3)

df_vereadores_partido = pd.merge(df_vereadores, df_partidos, how='left', on=['ano', 'cod_partido'])

df_vereadores_partido['cod_partido'].isnull().any()

df_vereadores_partido.head()

## Join emendas

df_emendas_vereadores.head()

df_emendas_vereadores_detalhado = pd.merge(df_emendas_vereadores, df_vereadores_partido, how='left', on=['ano', 'cod_vereador'])

df_emendas_vereadores_detalhado.head()

df_emendas_vereadores_detalhado['sigla'].isnull().any()

df_emendas_vereadores_detalhado.rename({'sigla' : 'sigla_partido', 'motivo' : 'motivo_emenda'}, axis=1, inplace=True)

df_emendas_vereadores_detalhado.groupby('apelido').count()['num_emenda'].nlargest(10)

df_emendas_vereadores_detalhado.groupby('sigla_partido').count()['num_emenda'].nlargest(10)

df_emendas_vereadores_detalhado['sigla_partido'].unique()

df_emendas_vereadores_detalhado.to_csv(os.path.join(DATA_FOLDER, 'df_merge_emendas_vereadores.csv'), index=False, sep=';')
