==> 05_join_emendas_vereador_and_loa.ipynb <==
## Join dados emendas Camara e Loa

Neste notebook fazemos o join entre o banco de dados que construímos com o planejamento das emendas 

import pandas as pd
import os

from config import DATA_FOLDER

df_loa = pd.read_csv(os.path.join(DATA_FOLDER, 'df_merge_loa_da.csv'), low_memory=False, sep=';')

df_loa.head()

df_emendas = pd.read_csv(os.path.join(DATA_FOLDER, 'df_merge_emendas_vereadores.csv'), low_memory=False, sep=';')

df_emendas.head()

df_emendas.rename(columns={'num_emenda': 'cod_emenda', 'ano' : 'exercicio'}, inplace=True)

df_emendas['cod_emenda'].astype(int).astype(str).str.len().max()

df_loa['cod_emenda'].astype(int).astype(str).str.len().max()

df_emendas['cod_emenda'] = df_emendas['cod_emenda'].astype(int).astype(str).str.zfill(4)
df_loa['cod_emenda'] = df_loa['cod_emenda'].astype(int).astype(str).str.zfill(4)


df_emendas['exercicio'].astype(int).astype(str).str.len().max()

df_loa['exercicio'].astype(int).astype(str).str.len().max()

df_emendas['exercicio'] = df_emendas['exercicio'].astype(int)
df_loa['exercicio'] = df_loa['exercicio'].astype(int)

df_emendas['exercicio'].unique()

df_emendas['exercicio'].unique()

df_merge = pd.merge(df_loa, df_emendas, how='left', on=['exercicio', 'cod_emenda'], indicator=True)

df_merge.head()

### Checagem de integridade

df_merge['_merge'].value_counts()

df_merge.columns

df_merge.groupby('tipo_regionalizacao').count()['exercicio']

# Selecionar as duas linhas duplicadas
duplicated_rows = df_merge[df_merge[['exercicio', 'cod_emenda', 'cod_orgao', 'cod_unidade', 'cod_proj_atividade', 'cod_vereador', 'cod_subprefeitura']].duplicated(keep=False)]

# Comparar as duas linhas
differences = duplicated_rows.iloc[0] != duplicated_rows.iloc[1]

# Exibir as colunas que diferem
columns_with_differences = differences[differences].index.tolist()
print(columns_with_differences)

duplicated_rows['valor_orcado']

Checamos na base do SOF e identificamos que o registro na verdade não é duplicado: ele possui fontes diferentes. Como não estamos taerbalhando com o detalhamento da fonte, vamos agregar os valores.

duplicated_rows

df_merge.loc[315, 'valor_orcado'] = df_merge.loc[315, 'valor_orcado'] + df_merge.loc[316, 'valor_orcado']

df_merge.drop(316, inplace=True)

df_merge.reset_index(inplace=True, drop=True)

df_merge[df_merge['tipo_regionalizacao']=='Despesa Regionalizável'].groupby('sigla_partido')['cod_emenda'].nunique().nlargest(5)

from config import DADOS_FINAL

df_merge.to_csv(DADOS_FINAL, sep=';', index=False)
